mrc/requirements.txt
transformers==4.6.0
torchtext==0.3.1
tensorboardX
nltk
matplotlib
seaborn
